feat(TDP-4104): preparation steps vanished (#699)
feat(TDP-4104): preparation steps vanished

* Remove the last steps without corresponding actions to avoid having all steps vanished
* Add a unitary test
* Do not remove the content (action)  of a step that is reused in another step

(cherry picked from commit ce88362db12c9cb6918a57a0ceaeb385ffd4fbde)

+++ dataprep-preparation/src/main/java/org/talend/dataprep/preparation/task/PreparationCleaner.java
@@ -21,6 +21,8 @@ import java.util.stream.Stream;
 
 import javax.annotation.Resource;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -37,6 +39,8 @@ import org.talend.dataprep.security.SecurityProxy;
 @Component
 @EnableScheduling
 public class PreparationCleaner {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(PreparationCleaner.class);
 
     @Autowired
     private PreparationRepository repository;
@@ -62,6 +66,7 @@ public class PreparationCleaner {
      *
      * @return The step ids
      */
+
     private Set<String> getPreparationStepIds() {
         return repository.list(Preparation.class) //
                 .flatMap(p -> p.getSteps().stream().map(Step::getId)) //
@@ -85,7 +90,6 @@ public class PreparationCleaner {
      */
     @Scheduled(fixedDelay = 60 * 60 * 1000, initialDelay = 60 * 60 * 1000) // Every hour
     public void removeOrphanSteps() {
-
         securityProxy.asTechnicalUser();
         try {
             getCurrentOrphanSteps().forEach(step -> {
@@ -95,9 +99,16 @@ public class PreparationCleaner {
                 repository.remove(stepToRemove);
 
                 // Remove actions linked to step
-                final PreparationActions preparationActionsToRemove = new PreparationActions();
-                preparationActionsToRemove.setId(step.getContent());
-                repository.remove(preparationActionsToRemove);
+                // if this step re-use an existing actions we don't delete the actions
+                boolean criterion = repository.exist(PersistentStep.class, "content" + "='" + step.getContent() + "'");
+                if (criterion) {
+                    LOGGER.info("Don't removing step content {} it still used by another step.", step.getContent());
+                } else {
+                    LOGGER.info("Removing step content {}.", step.getContent());
+                    final PreparationActions preparationActionsToRemove = new PreparationActions();
+                    preparationActionsToRemove.setId(step.getContent());
+                    repository.remove(preparationActionsToRemove);
+                }
 
                 // Remove metadata linked to step
                 final StepRowMetadata stepRowMetadataToRemove = new StepRowMetadata();
@@ -107,6 +118,5 @@ public class PreparationCleaner {
         } finally {
             securityProxy.releaseIdentity();
         }
-
     }
 }
